+++ src/Projects/Depths.Core/Constants/DScreenConstants.cs
@@ -2,8 +2,8 @@
 {
     internal static class DScreenConstants
     {
-        internal const ushort SCREEN_WIDTH = GAME_WIDTH * SCALE_FACTOR;
-        internal const ushort SCREEN_HEIGHT = GAME_HEIGHT * SCALE_FACTOR;
+        internal const int SCREEN_WIDTH = GAME_WIDTH * SCALE_FACTOR;
+        internal const int SCREEN_HEIGHT = GAME_HEIGHT * SCALE_FACTOR;
 
         internal const byte GAME_WIDTH = 84;
         internal const byte GAME_HEIGHT = 48;
